requirements.txt
python-telegram-bot==21.4
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
requests==2.32.3
python-dotenv==1.0.1
